+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
-## 1.0.0
+## 1.0.1 - 2020-03-03
+
+- Class WKT with static parse() method
+- Getters for WKT object parameters
+  
+## 1.0.0 - 2020-03-02
 
 - Initial version
 - It produces the same result as proj4js/wkt-parser for all the tests defined in proj4js/wkt-parser